Seigen: Rename `elastic_wave` -> `seigen`
