+++ seigen/data/marmhard.dat
