+++ seigen/marmousi.py
