+++ tests/LOH.1/LOH.1.py
@@ -1,8 +1,7 @@
 #!/usr/bin/env python
 
 from firedrake import *
-from elastic_wave.elastic import *
-from elastic_wave.helpers import *
+from seigen import *
 
 with timed_region('mesh setup'):
     mesh = Mesh("src/domain.msh")
